Consolidate fcrepo jms constants into constants file
