+++ services-camel/src/main/java/edu/unc/lib/dl/services/camel/solr/SolrIngestProcessor.java
@@ -16,7 +16,7 @@
 
 package edu.unc.lib.dl.services.camel.solr;
 
-import static edu.unc.lib.dl.services.camel.util.CdrFcrepoHeaders.FCREPO_RESOURCE_TYPE;
+import static edu.unc.lib.dl.services.camel.FcrepoJmsConstants.RESOURCE_TYPE;
 import static org.fcrepo.camel.FcrepoHeaders.FCREPO_URI;
 
 import java.util.ArrayList;
@@ -74,7 +74,7 @@ public class SolrIngestProcessor implements Processor {
 
         List<PID> targetPids = new ArrayList<>();
         PID targetPid = PIDs.get(fcrepoUri);
-        String resourceTypes = (String) in.getHeader(FCREPO_RESOURCE_TYPE);
+        String resourceTypes = (String) in.getHeader(RESOURCE_TYPE);
 
         // for binaries, need to index the file and work objects which contain it
         if (resourceTypes != null && resourceTypes.contains(Fcrepo4Repository.Binary.getURI())) {
