+++ services-camel/src/main/java/edu/unc/lib/dl/services/camel/util/CdrFcrepoHeaders.java
@@ -44,7 +44,5 @@ public abstract class CdrFcrepoHeaders {
 
     public static final String CdrEnhancementSet = "CdrEnhancementSet";
 
-    public static final String FCREPO_RESOURCE_TYPE = "org.fcrepo.jms.resourceType";
-
     public static final String CdrSolrUpdateAction = "CdrSolrUpdateAction";
 }
